refactor: restructure Drive module organization and implement barrel exports
