+++ src/views/Backups/components/DeviceList.tsx
@@ -5,7 +5,7 @@ import { DownloadSimple } from '@phosphor-icons/react';
 import Empty from '../../../app/shared/components/Empty/Empty';
 import notificationsService, { ToastType } from '../../../app/notifications/services/notifications.service';
 import { useTranslationContext } from '../../../app/i18n/provider/TranslationProvider';
-import { contextMenuBackupItems } from 'views/Drive/components/DriveExplorer/DriveExplorerList/DriveItemContextMenu';
+import { contextMenuBackupItems } from 'views/Drive/components/DriveExplorer/components';
 import { DriveFolderData } from '@internxt/sdk/dist/drive/storage/types';
 import { skinSkeleton } from 'app/shared/Skeleton';
 import { List } from '@internxt/ui';
