+++ src/views/Drive/components/DriveExplorer/components/DriveExplorerGridItem.scss
