+++ src/views/Shared/hooks/useSharedContextMenu.test.ts
@@ -6,9 +6,9 @@ import { renderHook } from '@testing-library/react';
 import { describe, expect, it, vi, beforeEach } from 'vitest';
 import useSharedContextMenu from './useSharedContextMenu';
 import { AdvancedSharedItem } from 'app/share/types';
-import * as DriveItemContextMenu from 'views/Drive/components/DriveExplorer/DriveExplorerList/DriveItemContextMenu';
+import * as DriveItemContextMenu from 'views/Drive/components/DriveExplorer/components/DriveItemContextMenu';
 
-vi.mock('views/Drive/components/DriveExplorer/DriveExplorerList/DriveItemContextMenu');
+vi.mock('views/Drive/components/DriveExplorer/components/DriveItemContextMenu');
 
 describe('useSharedContextMenu', () => {
   const mockActions = {
